feat: integrate top-up database saving and conditional history UI
- Implemented Firebase saving logic for user top-up transactions.
- Added conditional formatting in MyAdapter:
    - Green color and 'topup' icon for PLP (top-up) transactions.
    - Red color and 'send' icon for SYP (send) transactions.
- Fixed decimal truncation bug to ensure amounts display in 00.00 format.

+++ app/src/main/java/bankal_deir/com/transactions.kt
@@ -1,13 +1,13 @@
 package bankal_deir.com
 
 data class transactions(
-    val amount: Int = 0,
+    val amount: Double = 0.0,
     val date: String = "",
     val receiverWalletID: String = "",
     val senderUserId: String = "",
     val senderWalletID: String = "",
     val transactionNumber: String = ""
 ) {
-    constructor() : this(0, "", "", "", "", "")
+    constructor() : this(0.0, "", "", "", "", "")
 }
